Make IsReachable ignore all responses but a few critical ones
* Introduce proper clean-up and reconfiguration of NFTs to avoid leaks
