--- Explorer/Assets/DCL/SDKComponents/NFTShape/Imported/Resources.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 098653f35038f48428c863c581858af2
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
