+++ Explorer/Assets/DCL/SDKComponents/NFTShape/System/LoadCycleNftShapeSystem.cs
@@ -6,9 +6,11 @@ using DCL.ECSComponents;
 using DCL.SDKComponents.NFTShape.Component;
 using ECS.Abstract;
 using ECS.Prioritization.Components;
+using ECS.StreamableLoading.Common.Components;
 using ECS.StreamableLoading.NFTShapes;
 using ECS.StreamableLoading.NFTShapes.URNs;
 using ECS.Unity.Groups;
+using UnityEngine;
 using Promise = ECS.StreamableLoading.Common.AssetPromise<UnityEngine.Texture2D, ECS.StreamableLoading.NFTShapes.GetNFTShapeIntention>;
 
 namespace DCL.SDKComponents.NFTShape.System
@@ -29,6 +31,7 @@ namespace DCL.SDKComponents.NFTShape.System
         {
             StartQuery(World!);
             FinishAndApplyQuery(World!);
+            AbortLoadingIfComponentDeletedQuery(World);
         }
 
         [Query]
@@ -40,9 +43,18 @@ namespace DCL.SDKComponents.NFTShape.System
         }
 
         [Query]
+        [None(typeof(PBNftShape))]
+        private void AbortLoadingIfComponentDeleted(ref NFTLoadingComponent nftLoadingComponent)
+        {
+            nftLoadingComponent.Promise.ForgetLoading(World);
+
+            // Entity will be removed automatically with all remaining components
+        }
+
+        [Query]
         private void FinishAndApply(ref NFTLoadingComponent nftLoadingComponent, in NftShapeRendererComponent nftShapeRendererComponent)
         {
-            if (nftLoadingComponent.TryGetResult(World!, out var result))
+            if (!nftLoadingComponent.Promise.IsConsumed && nftLoadingComponent.Promise.TryConsume(World!, out StreamableLoadingResult<Texture2D> result))
             {
                 if (result.Succeeded)
                     nftShapeRendererComponent.PoolableComponent.Apply(result.Asset!);
