+++ Explorer/Assets/DCL/SDKComponents/NFTShape/Tests/NftShape.Tests.asmref
@@ -1,3 +1,3 @@
 {
-    "reference": "GUID:7213a1d35667f9b40be6aca8cbfcee23"
+    "reference": "GUID:da80994a355e49d5b84f91c0a84a721f"
 }
