--- Explorer/Assets/DCL/SDKComponents/NFTShape/Tests/NftShapeTest.cs
@@ -1,13 +1,0 @@
-using NUnit.Framework;
-
-namespace DCL.SDKComponents.NFTShape.Tests
-{
-    public class NftShapeTest
-    {
-        [Test]
-        public void EmptyTest()
-        {
-            //TODO test
-        }
-    }
-}
