+++ Explorer/Assets/DCL/Tests/Editor/DCL.EditMode.Tests.asmdef
@@ -63,7 +63,8 @@
         "GUID:702f733b4deb246808c6ce84d93b5c9c",
         "GUID:55c1fa1550cafa242ba23555f5a26b47",
         "GUID:2b89ae91f091424498ec65cf1ba6c557",
-        "GUID:ace653ac543d483ba8abee112a3ba2a6"
+        "GUID:ace653ac543d483ba8abee112a3ba2a6",
+        "GUID:c54a193e9535e43d8a25007ab5ee28b1"
     ],
     "includePlatforms": [
         "Editor"
