+++ Explorer/Assets/DCL/Tests/PlayMode/DCL.PlayMode.Tests.asmdef
@@ -56,7 +56,8 @@
         "GUID:e25ef972de004615a22937e739de2def",
         "GUID:1d2c76eb8b48e0b40940e8b31a679ce1",
         "GUID:875a5d5129614170bd769ed012c2eb3d",
-        "GUID:4725c02394ab4ce19f889e4e8001f989"
+        "GUID:4725c02394ab4ce19f889e4e8001f989",
+        "GUID:c54a193e9535e43d8a25007ab5ee28b1"
     ],
     "includePlatforms": [],
     "excludePlatforms": [
